pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "job-hunter"
version = "0.1.0"
description = "AI-powered job hunting assistant for Norwegian job market"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Nicolas Cook", email = "nicowc@gmail.com"}
]
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "sqlalchemy>=2.0.0",
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
    "pandas>=2.0.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "ratelimit>=2.2.1",
    "keyring>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
job-hunter = "src.cli:main"

[project.urls]
Homepage = "https://github.com/teztu/job_dashboard"
Repository = "https://github.com/teztu/job_dashboard"
Issues = "https://github.com/teztu/job_dashboard/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "E402",   # module level import not at top (required for Streamlit)
    "E712",   # comparison to True (required for SQLAlchemy filters)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.per-file-ignores]
"src/dashboard/**" = ["E402"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
